% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.constraints.R
\name{factor_covariance}
\alias{factor_covariance}
\title{Compute the square-root and inverse square-root of a non-negative definite matrix.}
\usage{
factor_covariance(S, rank = NA)
}
\arguments{
\item{S}{matrix}

\item{rank}{rank of svd}
}
\description{
Compute the square-root and inverse square-root of a non-negative definite matrix.
}

