% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.constraints.R
\name{thresh2constraints}
\alias{thresh2constraints}
\title{Translate between coordinate thresholds and affine constraints}
\usage{
thresh2constraints(d, lower = rep(-Inf, d), upper = rep(Inf, d))
}
\arguments{
\item{d}{dimension of vector}

\item{lower}{1 or d-dim lower constraints}

\item{upper}{1 or d-dim upper constraints}
}
\description{
\code{thresh2constraints} translates lower and upper constraints 
on coordinates into linear and offset constraints (A*Z <= B).
lower and upper can have -Inf or Inf coordinates.
}

