% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.constraints.R
\name{whiten_constraint}
\alias{whiten_constraint}
\title{Transform non-iid problem into iid problem}
\usage{
whiten_constraint(linear_part, offset, mean_param, covariance)
}
\arguments{
\item{linear_part}{matrix, linear part of constraints}

\item{offset}{vector, bias of constraints}

\item{mean_param}{vector of unconditional means}

\item{covariance}{vector of unconditional covariance}
}
\value{
new \code{linear_part} and \code{offset} for 0-mean iid covariance problem, 
and functions that map between the two problems.
}
\description{
Transform non-iid problem into iid problem
}

